add till step 9
At this stage, I am envisioning to provide only 1 yaml file to read initial location of input files, add 1 option to either use ortho in stereo processing at all or not, compute slope of area from initail refdem, if slope is less than < 15 degrees, go for Nuth&Kaab, otherwise go for ICP alignment. Also, would be good to delete the intermediate files and keep only the final products which might be needed if this is being tested on a very nasty source of input (maybe this can be made default, with option in yaml file to preserve intermediate files if needed).